pyproject.toml
[project]
name = "dotdocs"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "chromadb>=1.0.16",
    "clerk-backend-api>=3.1.11",
    "fastapi>=0.116.1",
    "fastembed>=0.7.1",
    "httpx>=0.28.1",
    "openai>=1.99.6",
    "passlib[bcrypt]>=1.7.4",
    "pinecone>=7.3.0",
    "pip>=25.2",
    "pydantic-ai>=0.6.2",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-jose[cryptography]>=3.5.0",
    "qdrant-client>=1.15.1",
    "scrapy>=2.13.3",
    "spacy>=3.8.7",
    "structlog>=25.4.0",
    "typer>=0.16.0",
    "uvicorn[standard]>=0.35.0",
]
